Fix browser rendering issue: pure color tiles are rendered black
The “pure color tile” optimization finds the pure-color tiles and
then draws the tiles simply with the color (otherwise the texture).

But the RGBA color pointer type is char, which may overflow when
the value is larger than 127. In such case, the tiles in question
are rendered black, because of the wrong color. This bug is fixed
by defining the pointer as unsigned char.

Change-Id: I5c7214423fa6961e707bde72f0e18c610374745d
Author: Yuyang Du <yuyang.du@intel.com>
Signed-off-by: Yuyang Du <yuyang.du@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>